Remove all unused variables and functions
